confirm pane refactor
