+++ web/src/components/common/adminComponents/ShowUsers.jsx
@@ -9,6 +9,7 @@ import Backend from '../../common/Backend'
 import CircularProgress from 'material-ui/CircularProgress'
 import {Card, CardHeader, CardText} from 'material-ui/Card'
 import TextField from 'material-ui/TextField'
+import ConfirmationPane from '../../common/ConfirmationPane'
 
 
 export default class ShowUsers extends Component {
@@ -27,6 +28,7 @@ export default class ShowUsers extends Component {
     this.handleClose = this.handleClose.bind(this)
     this.handleChange = this.handleChange.bind(this)
     this.handleSetAdmin = this.handleSetAdmin.bind(this)
+    this.intentionToDeleteUser = this.intentionToDeleteUser.bind(this)
   }
 
   handleClose(){
@@ -61,6 +63,7 @@ export default class ShowUsers extends Component {
   }
 
   componentWillReceiveProps(props){
+    alert("fd")
     this.setState({
       loading : props.loading,
       open : false,
@@ -97,11 +100,17 @@ export default class ShowUsers extends Component {
     })
   }
 
-
-  handleDelete(userID){
+  intentionToDeleteUser(userID){
     this.setState({
-        open : true,
-        userID : userID})
+      open : true,
+      userID : userID})
+  }
+
+  handleDelete(){
+    Backend.deleteUser(this.state.userID).then(function confirm(){
+      this.setState({open:false})
+      this.props.handleUserUpdate("User " + this.state.userID + " deleted")
+    }.bind(this))
   }
 
   render(){
@@ -182,7 +191,7 @@ export default class ShowUsers extends Component {
       width : 60,
       style : {textAlign : "center"},
       Cell : props => <IconButton
-            onClick = {()=>this.handleDelete(props.value.email)}
+            onClick = {()=>this.intentionToDeleteUser(props.value.email)}
             touch = {true}
             style = {styles.largeIcon}
             iconStyle = {{fontSize : '24px'}}
@@ -195,9 +204,11 @@ export default class ShowUsers extends Component {
     return (
       <div className = "padding">
         <ConfirmationPane open = {this.state.open}
-                          userID = {this.state.userID}
-                          handleUserUpdate = {this.props.handleUserUpdate}
-                          handleClose = {this.handleClose}/>
+                          handleClose = {this.handleClose}
+                          title = {"Do you want to delete user " + this.state.userID}
+                          confirmationLabel = {"Delete User"}
+                          confirmAction = {this.handleDelete}
+        />
         <div className = "container" style = {{display : (this.state.loading ? "block" : "none")}}>
           <div className = "header"><CircularProgress size = {80} thickness = {5} /></div>
         </div>
@@ -258,48 +269,6 @@ export default class ShowUsers extends Component {
           </div>
         </div>
       </div>
-    )
-  }
-}
-
-class ConfirmationPane extends Component {
-  handleDelete = () =>{
-    var text = "You can not delete the admin user"
-    text = "User " + this.props.userID + " deleted"
-    Backend.deleteUser(this.props.userID).then(
-      this.props.handleUserUpdate(text)
-    )
-  }
-
-
-  componentWillReceiveProps(props){
-    this.setState({open : props.open})
-  }
-
-  render() {
-    const actions = [
-      <RaisedButton
-        label = "Cancel"
-        primary = {true}
-        onTouchTap = {this.props.handleClose}
-        />,
-      <RaisedButton
-        label = "DELETE USER"
-        primary = {true}
-        onTouchTap = {this.handleDelete}
-        style = {{marginLeft : 20}}
-        />,
-    ]
-
-    return (
-      <Dialog
-        title = {"Do you want to delete user "+ this.props.userID}
-        actions = {actions}
-        modal = {false}
-        open = {this.props.open}
-        onRequestClose = {this.props.handleClose}
-        >
-      </Dialog>
     )
   }
 }
